// src/generated/spec-enums.ts
export const RULESET_TARGETS = ["branch", "tag", "push"] as const;

export const RULESET_ENFORCEMENTS = ["disabled", "active", "evaluate"] as const;
// ...
export const BypassActorConfig = z
  .looseObject({
    actor_id: z
      .int()
      .nullable()
      .describe(
        [
          "The ID of the actor that can bypass a ruleset. Required for `Integration`,",
          "`RepositoryRole`, `Team`, and `User` actor types. If `actor_type` is",
          "`OrganizationAdmin`, `actor_id` is ignored. If `actor_type` is `DeployKey`, this",
          "should be null. `OrganizationAdmin` is not applicable for personal repositories.",
        ].join(" "),
      )
      .optional(),
    actor_type: z
      .enum(["Integration", "OrganizationAdmin", "RepositoryRole", "Team", "DeployKey", "User"])
      .describe("The type of actor that can bypass a ruleset."),
    bypass_mode: z
      .enum(["always", "pull_request", "exempt"])
      .describe(
        [
          "When the specified actor can bypass the ruleset. `pull_request` means that an",
          "actor can only bypass rules on pull requests. `pull_request` is not applicable",
          "for the `DeployKey` actor type. Also, `pull_request` is only applicable to",
          "branch rulesets. When `bypass_mode` is `exempt`, rules will not be run for that",
          "actor and a bypass audit entry will not be created.",
        ].join(" "),
      )
      .optional(),
  })
  .meta({
    id: "BypassActorConfig",
    description: "An actor that can bypass rules in a ruleset",
  });

export const RulesetConditions = z
  .looseObject({
    ref_name: z
      .looseObject({
        include: z
          .array(z.string())
          .describe(
            [
              "Array of ref names or patterns to include. One of these patterns must match for",
              "the condition to pass. Also accepts `~DEFAULT_BRANCH` to include the default",
              "branch or `~ALL` to include all branches.",
            ].join(" "),
          )
          .optional(),
        exclude: z
          .array(z.string())
          .describe(
            [
              "Array of ref names or patterns to exclude. The condition will not pass if any of",
              "these patterns match.",
            ].join(" "),
          )
          .optional(),
      })
      .optional(),
  })
  .describe("Parameters for a repository ruleset ref name condition");